requirements.txt
python-dotenv~=0.19.2
fastapi~=0.70.0
uvicorn